allow 32 and 64 bit builds of windows application
build.xml: specify correct classpath to use ant-javafx.jar from installed Java location, seperate build, package, bundle tasks for building 64 bits (allows unique names for .msi file and allows for future expansion - is WET code, however, which could make maintenance harder).
win-build.sh: break apart 32/64 bit build processes so we can specify correct ant tasks, set proper env var that will later be read by wxs file, read version from text file
sensorconnector.wxs: read registry to determine 32/64 bit build, act accordingly to configure build correctly based on if it is 32/64 bit version.
version.txt: set version in text file for reading later
[#159744652]

+++ script/win-build.sh
@@ -1,8 +1,22 @@
 #!/bin/sh
 
-export SC_APP_VERSION=`cat version.txt`
-export SC_ARCH_BITS=32
+# Usage: With no arguments builds 32-bit and 64-bit packages
+#        Use "-32" or "-64" argument to build 32-bit/64-bit package only
 
-rm -rf dist
-ant win-bundle && mv dist/SensorConnector-*.msi . && mv dist/SensorConnector-*.exe .
-rm *unsigned*
+export SC_APP_VERSION=`cat version.txt`
+
+# build 32-bit package
+if [[ ("$1" == "-32") || ($# -eq 0) ]]; then
+	export SC_ARCH_BITS=32
+	rm -rf dist
+	ant win-bundle && mv dist/SensorConnector-*.msi . && mv dist/SensorConnector-*.exe .
+	rm *unsigned*
+fi
+
+# build 64-bit package
+if [[ ("$1" == "-64") ]]; then
+	export SC_ARCH_BITS=64
+	rm -rf dist
+	ant win-bundle-x64 && mv dist/SensorConnector-x64-*.msi . && mv dist/SensorConnector-x64-*.exe .
+	rm *unsigned*
+fi
